education-agent/pyproject.toml
[project]
name = "education-agent"
version = "0.1.0"
description = "Education planner agent powered by LangGraph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.7",
    "langgraph-cli[inmem]>=0.4.2",
    "langsmith>=0.4.28",
    "openai[aiohttp]>=1.107.3",
    "python-dotenv>=1.1.1",
    "streamlit>=1.48.1",
    "tiktoken>=0.11.0",
]

[dependency-groups]
dev = ["ipykernel>=6.30.1", "pytest>=8.0.0", "pytest-asyncio>=0.24.0"]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.30.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
